Split CodeMiner into smaller methods
