+++ modules/pipeline/github-maven-project-miner/build.gradle
