+++ modules/pipeline/github-maven-project-miner/gradle.properties
