+++ modules/pipeline/src/main/kotlin/org/cafejojo/schaapi/pipeline/CodeMiner.kt
@@ -7,11 +7,11 @@ import org.cafejojo.schaapi.models.Project
  */
 interface CodeMiner {
     /**
-     * Mine projects which (likely) depend on a software library with the given group id, artifact id and version.
+     * Mine projects which (likely) depend on the library with the given group id, artifact id and version (number).
      *
      * @param groupId group id of library
      * @param artifactId artifact id of library
-     * @param version version of library
+     * @param version version (number) of library
      * @return list of projects using library
      */
     fun mine(groupId: String, artifactId: String, version: String): List<Project>
