+++ settings.gradle
@@ -19,7 +19,7 @@ includeModule "models", "jimple-library-usage-graph"
 includeModule "models", "java-project"
 
 includeModule "pipeline"
-includeModule "pipeline", "github-miner"
+includeModule "pipeline", "github-maven-project-miner"
 includeModule "pipeline", "java-jar-project-compiler"
 includeModule "pipeline", "java-maven-project-compiler"
 includeModule "pipeline", "jimple-evosuite-test-generator"
